requirements.txt
blinker==1.9.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
Flask==3.1.2
google-auth==2.47.0
google-auth-oauthlib==1.2.4
gspread==6.2.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
numpy==2.4.1
oauthlib==3.3.1
pandas==3.0.0
pyasn1==0.6.2
pyasn1_modules==0.4.2
python-dateutil==2.9.0.post0
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
urllib3==2.6.3
Werkzeug==3.1.5
